requirements.txt
black
flake8
flask
lxml
markdown
mkdocs
pygments
pymdown-extensions
pymongo
pyyaml
